ai_microservice/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
llama-cpp-python==0.2.59
python-multipart==0.0.20
